pyproject.toml
[project]
name = "engineering-standards"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "jinja2>=3.1.6",
    "pyyaml>=6.0.3",
]
